backend/lambda/requirements.txt
boto3==1.35.0
botocore==1.35.0
pydantic==2.6.4
python-dateutil==2.9.0
typing-extensions==4.9.0
PyJWT[crypto]==2.8.0
cryptography==42.0.5
requests==2.31.0
pytest==8.2.0
pytest-mock==3.14.0

